修改Bestman相关命名与接口顺序等信息，删除不必要的代码
