+++ RoboticsToolBox/__init__.py
@@ -1,2 +1,2 @@
-from .Bestman import Bestman
+from .Bestman_sim import Bestman_sim
 from .Pose import Pose
